rename modelmilestone to milestone only
